cleaned up code base split function files out to more focused files
