+++ fcn_Importing.py
@@ -1,0 +1,64 @@
+
+import pandas as pd
+import ctypes
+import cufflinks as cf
+#IMPORT USER DEFINED GLOBAL VARIABLES 
+import config
+
+
+def xlsxReader_Monthly(xls_file_path): 
+    """reads a given file (xls_file_path) and returns a list of DataFrames split into months
+    Access said dataframe via indexing
+    ie, JAN = 0
+        FEB = 1
+        ...
+        DEC = 11
+    """
+    ### STEP 1 -  read the data without index files
+    data = pd.read_excel(xls_file_path, parse_dates = True, index_col = None) #reads entire df and parses dates without creating an index
+    
+    months = [g for n, g in data.groupby(pd.Grouper(key='Interval End',freq='M'))] #splits it into months
+        # is a list, so just access each list as an index (ie, JAN = 0, FEB = 1)
+        # https://stackoverflow.com/a/49491178/13181119
+    
+    return months
+
+def Extension_Checker(file_name_to_check):
+    """ used to check if the extension is a xls(x) or a csv, or returns an error if not"""
+    if file_name_to_check.endswith('.xls') or file_name_to_check.endswith('.xlsx'): #open via xls reader
+        try: 
+            read_file = xlsxReader_Monthly(file_name_to_check)
+        except AttributeError: 
+            pass
+    elif file_name_to_check.endswith('.csv') or file_name_to_check.endswith('.xlsx'): #open via csv reader
+        print('PLEASE ENSURE FILE IS A \'XLSX\' ONLY') #make this a bit better later
+    else: 
+        print('ERROR') #make this a bit better later
+    
+    return read_file
+
+def Data_Consistency_Checker(Dataframe_to_check):
+    """Used to check if the dataframe is fully populated, and returns a interpolated dataframe is necessary"""
+    
+    ### STEP 1: Ensure all data is consistent (ie, exists)
+    for x in  range(0, len(Dataframe_to_check)):
+        if Dataframe_to_check[x].isnull().any().any(): #returns TRUE if NaN exists in teh dataframe
+            Dataframe_to_check[x] = intervalResampler(Dataframe_to_check[x])
+    
+    return Dataframe_to_check
+
+def intervalResampler(input_df, chosen_interval = 30):
+    """
+    function to change and interpolate dataframe to a given interval 
+    """
+    
+    Index_Name = 'Interval End'
+    # resampling_df = input_df #each month
+    input_df.set_index(Index_Name, inplace = True) #set the datetime index 
+    resampledDF = input_df.resample('30T').interpolate(method = 'polynomial', order = 2) #interpolate the hourly interval data to 30 mins via linear interpolation   , inplace = True
+    resampledDF.reset_index(inplace = True)
+    
+    return resampledDF
+
+
+
