+++ myproject/app.py
@@ -19,8 +19,11 @@ import base64
 import PySimpleGUI as sg
 import cufflinks as cf
 #USER CREATED FUNCTIONS 
-from fcn_UTILS import dataJoiner, xlsxReader_Monthly, Extension_Checker, intervalResampler, Data_Consistency_Checker, CopyCat, character_removal, dataframe_chooser, Mbox, dash_solar_plotter, load_shifter_average, solar_extractor_adder, dataframe_list_generator, dataframe_compactor, load_shifter_scratch, load_shifter_long_list
 from fcn_Averages import DailyAverage, WeeklyAverage, MonthToDaySum, ConsumptionSummer
+from fcn_plotting import character_removal, dataframe_chooser, Mbox, dash_solar_plotter
+from fcn_Importing import xlsxReader_Monthly, Extension_Checker, Data_Consistency_Checker, intervalResampler
+from fcn_loadshifting import load_shifter_average, load_shifter_long_list, solar_extractor_adder
+from fcn_UTILS import dataJoiner, CopyCat, dataframe_list_generator, dataframe_compactor
 #IMPORT USER DEFINED GLOBAL VARIABLES 
 import config
 
@@ -93,108 +96,11 @@ config.Weekly_Interval_Data = WeeklyAverage(Checked_Interval_Data_0)
 ## STEP 5: Calculate Daily Averages
 config.Daily_Interval_Data = DailyAverage(Checked_Interval_Data_1)
 
-
 ## STEP 6: Calculate summation of energy used (Yearly, monthly, weekly, daily)
 config.Monthly_Sum = ConsumptionSummer(Checked_Interval_Data_1) #total average month (x12 months)
 
 #########////////////////////////\\\\\\\\\\\\\\\\\\\\#################
 print('succesfully loaded and did the backend stuff')
-
-######## CREATING LOAD SHIFTING ################# TEMP #############
-
-#     #### VARS ####
-
-
-# chosen_site = 'Faithful Street WANGARATTA - kWh Consumption'
-# month = 0
-
-
-# ### STEP 1 - narrow down dataframe to chosen site 
-# site_to_plot_raw = dataframe_chooser(config.Daily_Interval_Data, chosen_site) #dynamically create dataframes to plot 12x months on top of each other for the selected site
-# #the above returns a 12x? dataframe. need to convert it to a list of dataframes 
-
-# ### STEP 2 - Convert the dataframe for a list of dataframes 
-# site_to_plot_list = dataframe_list_generator(non_list_dataframe = site_to_plot_raw) #converts the above into a list of 1x month per list entry 
-
-# # ### STEP 3 - extract solar from the original dataframe, and add it to each list 
-# site_to_plot_solar_added = solar_extractor_adder(single_site = site_to_plot_list, all_sites = config.Daily_Interval_Data) #adds the respective monthly solar to the respective month (in the list)
-
-# dataframe_to_shift = site_to_plot_solar_added
-# ### FUNCTION STARTS HERE 
-
-# shifted_site_consumption = [] 
-# inverter = -1 #used to invert the shifted hours when adding a negative, you minus - NEEDED
-# ### TESTING PURPOSES
-# value_to_shift = 50 # FOR TESTING 
-# ### TESTING PURPOSES
-
-# ### STEP 1 - convert value_to_shift into % (ie, 20 = 0.2)
-# value_to_shift_percentage = value_to_shift/100 
-
-# #isolate Dataframe to work on
-# single_df_site = dataframe_to_shift[month]
-
-# #create new dataframe only where consumption > PV availability        IE< NON SOLAR EXCESS HOURS / OUTSIDE SOLAR HOURS
-# no_solar_hours_consumption = single_df_site.iloc[:,0].where(single_df_site.iloc[:,0] > single_df_site['Excess Solar Generation (Total)']) #change name it OUTSIDE
-
-# #create new dataframe only where consumption < PV availability        IE< SOLAR EXCESS HOURS / INSIDE SOLAR HOURS
-# solar_hours_consumption = single_df_site.iloc[:,0].where(single_df_site.iloc[:,0] < single_df_site['Excess Solar Generation (Total)']) #change name to INSIDE
-
-# # sum the total excess solar hours 
-# solar_summed = solar_hours_consumption.sum() #total (summed) SOLAR GENERATION to shift
-
-# # determine solar ratio by dividing half hourly solar generation by total solar generation 
-# solar_hours_consumption_ratio = solar_hours_consumption/solar_summed #add this to the summed NON SOLAR EXCESS load #CHANGE NAME TO INSIDE
-
-# ## shift NON EXCESS SOLAR hours by value_to_shift_percentage,           value is negative, as we want to TAKE AWAY these hours
-# no_solar_hours_consumption_scaled = no_solar_hours_consumption*value_to_shift_percentage*inverter #multiple to get smaller number (ie, number to add to original dataframe) #change name to OUTSIDE
-
-# ### STEP 4 - sum total SHIFTED HOURS (ie, in NO EXCESS SOLAR)
-# summed = no_solar_hours_consumption_scaled.sum() #total kWh in NON SOLAR HOURS to shift
-# summed_positive = summed*inverter #to give a positive number for dividing 
-
-# #create a dataframe with consumption to ADD to each interval INSIDE SOLAR HOURS
-# scaled_inside_solar_hours_consumption = solar_hours_consumption_ratio*summed_positive #this is what we need to ADD to INSIDE SOLAR HOURS
-
-# #adding the scaled shifted consumption to the original solar hours 
-# shifted_inside_solar_hours = solar_hours_consumption + scaled_inside_solar_hours_consumption 
-
-# #take away the shifted load from the original 
-# shifted_outside_solar_hours = no_solar_hours_consumption + no_solar_hours_consumption_scaled
-
-# #combine the two dataframe 
-# shifted_inside_solar_hours.update(shifted_outside_solar_hours)
-
-
-
-
-# ### STEP 7 - recreate new dataframe by subtracting SHIFTED HOURS and adding EXCESS SOLAR HOUR to each hour in the original dataframe  
-# shifted_solar_hours_summed = solar_hours_consumption_ratio*summed #this is what we need to ADD to INSIDE SOLAR HOURS
-
-
-# ## add the INSIDE SOLAR SHIFTED HOURS to the original INSIDE SOLAR hours
-# FINAL_INSIDE_SOLAR_HOURS = shifted_solar_hours_summed+shifted_dataframe_solar_hours
-
-# shifted_dataframe_no_solar_hours.update(FINAL_INSIDE_SOLAR_HOURS) #join the two dataframes as one
-
-# shifted_site_consumption.append(shifted_dataframe_no_solar_hours) #for punching through each month - disregard for dev
-
-
-
-# #convert list of dataframe to single dataframe 
-# single_site_dataframe = dataframe_compactor(dataframes_to_compact = shifted_site_consumption) #converts the list of dataframes to a single dataframe 
-
-
-
-
-
-
-
-
-
-
-
-
 
 ########## VARS SPECIFICALLY FOR DASH  ###############
 names = list(config.Daily_Interval_Data[0].columns) #get the names of the column, assuming every name is the same across each dataframe in the list
@@ -202,20 +108,19 @@ chosen_site = '' #to make this VAR global
 image_filename = str(os.getcwd()) + '\\assets\\NEW_LOGO.jpg' # replace with your own image
 encoded_image = base64.b64encode(open(image_filename, 'rb').read())
 
-
+#create excess solar plots for DASH
 if Solar_Exists: #only make this if the solar data has been uploaded
     solar_figure_summed = dash_solar_plotter(df_to_plot = config.Daily_Interval_Data, plot_type = 'bar' ) #make fancy figure 
     solar_figure_line = dash_solar_plotter(df_to_plot = config.Daily_Interval_Data, plot_type = 'line' ) #make fancy figure 
 
 
-
-### DASH ###
+##################////////////////// DASH \\\\\\\\\\\\\\\\\\################
 
 external_stylesheets = ['https://codepen.io/chriddyp/pen/bWLwgP.css']
 
 app = dash.Dash(__name__, external_stylesheets=external_stylesheets)
 
-app.layout = html.Div([ ### MINIMAL LAYOUT###
+app.layout = html.Div([ ### LAYOUT FOR TABS - ACTUAL LAYOUT IS DEFINED INSIDE TAB CALLBSCKS###
     html.Img(src='data:image/png;base64,{}'.format(encoded_image.decode())), #DISPLAY the NEW LOGO 
     dcc.Tabs(id='tabs-example', value='tab-1', children=[ #DISPLAY TABS
         dcc.Tab(label='Load Data', value='tab-1'),
